requirements-ai.txt
openai>=1.0.0
langchain>=0.1.0
pydantic>=2.0.0
fastapi>=0.109.0
python-docx>=1.1.0
aiofiles>=23.2.1
PyPDF2>=3.0.0
python-multipart>=0.0.6
tiktoken>=0.5.2
uvicorn[standard]>=0.27.0
